requirements.txt
numpy
opencv-contrib-python
pandas
Pillow
pyttsx3
